+++ app/src/main/java/com/example/watchlist/Watch.kt
@@ -3,7 +3,7 @@ package com.example.watchlist
 import android.net.Uri
 
 public class Watch(
-        val Id : Int = 0,
+        val Id : Long = 0,
         var Title :String = "",
         var Content:String = "",
         var Date:String = "",
