+++ app/src/main/java/com/example/watchlist/WatchViewActivity.kt
@@ -15,7 +15,7 @@ class WatchViewActivity : AppCompatActivity() {
         window.setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN)
         setContentView(R.layout.activity_watch_view)
 
-        val id = intent.getIntExtra("id", 0)
+        val id = intent.getLongExtra("id", 0)
         val watch = watchListRepository.getWatchListById(id)
         val deteleButton = findViewById<Button>(R.id.DeleteWatchList)
         val updateButton = findViewById<Button>(R.id.updateWatchList)
